MEOW

+++ app.py
@@ -1,8 +1,8 @@
-from fastapi import FastAPI, HTTPException, Depends
+from fastapi import FastAPI, HTTPException, Depends, Header
 from fastapi.responses import RedirectResponse
 from pydantic import BaseModel, HttpUrl
 from sqlalchemy.orm import Session
-import random, string
+import random, string, secrets, bcrypt
 
 from database import Base, engine, get_db
 from models import URL
@@ -17,19 +17,31 @@ class URLRequest(BaseModel):
 def make_code() -> str:
     return "".join(random.choices(string.ascii_letters + string.digits, k=6))
 
+def make_api_key():
+    return secrets.token_urlsafe(32)
+
+def hash_api_key(api_key: str) -> str:
+    return bcrypt.hashpw(api_key.encode(), bcrypt.gensalt()).decode()
+
+def verify_api_key(api_key: str, hashed: str) -> bool:
+    return bcrypt.checkpw(api_key.encode(), hashed.encode())
+
 @app.post("/shorten")
 def shorten(request: URLRequest, db : Session = Depends(get_db)) -> dict:
     code = make_code()
-    entry = URL(code=code, destination=str(request.url))
+    api_key = make_api_key()
+    entry = URL(code=code, destination=str(request.url), api_key=hash_api_key(api_key))
     db.add(entry)
     db.commit()
-    return {"short_url": f"http://localhost:8000/{code}"}
+    return {"short_url": f"http://localhost:8000/{code}", "api_key": api_key}
 
 @app.delete("/delete/{code}")
-def delete(code: str, db : Session = Depends(get_db)):
+def delete(code: str, x_api_key: str = Header(), db : Session = Depends(get_db)):
     entry = db.query(URL).filter(URL.code == code).first()
     if not entry:
         raise HTTPException(status_code=404, detail="Link not found")
+    if not verify_api_key(x_api_key, entry.api_key):
+        raise HTTPException(status_code=403, detail="Invalid API key")
     db.delete(entry)
     db.commit()
     return {"message": f"Redirect {code} succesfully deleted"}
